+++ app/controllers/application_controller.rb
@@ -150,6 +150,9 @@ class ApplicationController < ActionController::Base
     session[:user_id] = user.id
     session[:user_login_token] = user.login_token
     session[:remember_me] = remember_me
+    # Intentionally allow to fail silently so that we don't have to care whether
+    # sign in recording is enabled.
+    user.sign_ins.create(ip: user_ip, country: country_from_ip)
   end
 
   # Logout form